# food_db/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Macro_vision",
        "title": "Macro_vision backend",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Verify user’s login credentials (username and password), then generate a session and return the session token back to user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "login",
                "parameters": [
                    {
                        "description": "login parameters",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginResponse"
                        }
                    },
                    "404": {
                        "description": "User not found in db",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Invalid Credentials",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "description": "Verify the received token and perform session clean up.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "logout",
                "parameters": [
                    {
                        "type": "string",
                        "example": "f3d9c4e6a7b1ce204fa8d5b39e181f9b3e2c1d7fbe4490d6732eab5c4fd7c92e",
                        "description": "256bit random token",
                        "name": "Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "Receive the sign up information, validate username (check if it exists), password (has to be lower than 71 byte for bcrypt hashing), and email (correct format). Then generate the session and return the token to the user (auto login)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "signup",
                "parameters": [
                    {
                        "description": "signup parameters",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.SignupParam"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Account Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.SignupResponse"
                        }
                    },
                    "401": {
                        "description": "Password too long (over 71 bytes)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "user alr in db",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/dishes/search": {
            "get": {
                "description": "Search dishes by name, course list, and ingredient count range. Results are paginated and include total match count.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dishes"
                ],
                "summary": "Search dishes with filters and pagination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query for dish name",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Filter by course (can be repeated)",
                        "name": "course",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum number of ingredients",
                        "name": "min_ingredients",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of ingredients",
                        "name": "max_ingredients",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results per page (default: 12)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (1-based)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "schema": {
                            "$ref": "#/definitions/handlers.SearchDishesResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/dishes/{dish_id}": {
            "get": {
                "description": "Retrieve a single dish by its UUID. If the dish does not exist, an empty object is returned.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dishes"
                ],
                "summary": "Get dish details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dish UUID",
                        "name": "dish_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dish details",
                        "schema": {
                            "$ref": "#/definitions/database.Dish"
                        }
                    },
                    "400": {
                        "description": "Invalid dish ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "auth.LoginParam": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "11111"
                },
                "username": {
                    "type": "string",
                    "example": "ranto"
                }
            }
        },
        "auth.SignupParam": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string",
                    "example": "phantom"
                },
                "email": {
                    "type": "string",
                    "example": "ranto@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "1233445"
                },
                "username": {
                    "type": "string",
                    "example": "ranto"
                }
            }
        },
        "database.Dish": {
            "type": "object",
            "properties": {
                "alt_name": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "course": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dish_id": {
                    "type": "string"
                },
                "dish_name": {
                    "type": "string"
                },
                "full_recipe": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
...
        "handlers.Dish": {
            "type": "object",
            "properties": {
                "alt_name": {
                    "type": "string"
                },
                "course": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dish_id": {
                    "type": "string"
                },
                "dish_name": {
                    "type": "string"
                },
                "full_recipe": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                }
            }
        },
...
        "handlers.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "f3d1b8cd4f29e6a173c0dd9d84b3af2d2e19a74ccf8b0e57a3c4d90e8f12b7ac"
                }
            }
        },
...
        "handlers.SearchDishesResponse": {
            "type": "object",
            "properties": {
                "dishes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.Dish"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_results": {
                    "type": "integer"
                }
            }
        },
        "handlers.SignupResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "f3d1b8cd4f29e6a173c0dd9d84b3af2d2e19a74ccf8b0e57a3c4d90e8f12b7ac"
                }
            }
        },
...
        "sql.NullString": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    }
}